LNURL-withdraw: move non-E2E tests to sdk-common
