+++ libs/sdk-common/src/lnurl/specs/withdraw.rs
@@ -162,3 +162,35 @@ pub mod model {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use anyhow::Result;
+
+    use crate::lnurl::tests::rand_string;
+    use crate::prelude::*;
+
+    fn get_test_withdraw_req_data(min_sat: u64, max_sat: u64) -> LnUrlWithdrawRequestData {
+        LnUrlWithdrawRequestData {
+            min_withdrawable: min_sat * 1000,
+            max_withdrawable: max_sat * 1000,
+            k1: rand_string(10),
+            default_description: "test description".into(),
+            callback: "http://localhost:8080/callback".into(),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_lnurl_withdraw_validate_amount_failure() -> Result<()> {
+        let invoice_str = "lnbc110n1p38q3gtpp5ypz09jrd8p993snjwnm68cph4ftwp22le34xd4r8ftspwshxhmnsdqqxqyjw5qcqpxsp5htlg8ydpywvsa7h3u4hdn77ehs4z4e844em0apjyvmqfkzqhhd2q9qgsqqqyssqszpxzxt9uuqzymr7zxcdccj5g69s8q7zzjs7sgxn9ejhnvdh6gqjcy22mss2yexunagm5r2gqczh8k24cwrqml3njskm548aruhpwssq9nvrvz";
+        let invoice = crate::invoice::parse_invoice(invoice_str)?;
+        let withdraw_req = get_test_withdraw_req_data(0, 1);
+
+        // Fail validation before even calling the endpoint (no mock needed)
+        assert!(validate_lnurl_withdraw(withdraw_req, invoice)
+            .await
+            .is_err());
+
+        Ok(())
+    }
+}
