+++ libs/sdk-core/src/lnurl/withdraw.rs
@@ -63,20 +63,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_lnurl_withdraw_validate_amount_failure() -> Result<()> {
-        let invoice_str = "lnbc110n1p38q3gtpp5ypz09jrd8p993snjwnm68cph4ftwp22le34xd4r8ftspwshxhmnsdqqxqyjw5qcqpxsp5htlg8ydpywvsa7h3u4hdn77ehs4z4e844em0apjyvmqfkzqhhd2q9qgsqqqyssqszpxzxt9uuqzymr7zxcdccj5g69s8q7zzjs7sgxn9ejhnvdh6gqjcy22mss2yexunagm5r2gqczh8k24cwrqml3njskm548aruhpwssq9nvrvz";
-        let invoice = crate::invoice::parse_invoice(invoice_str)?;
-        let withdraw_req = get_test_withdraw_req_data(0, 1);
-
-        // Fail validation before even calling the endpoint (no mock needed)
-        assert!(validate_lnurl_withdraw(withdraw_req, invoice)
-            .await
-            .is_err());
-
-        Ok(())
-    }
-
-    #[tokio::test]
     async fn test_lnurl_withdraw_endpoint_failure() -> Result<()> {
         let invoice_str = "lnbc110n1p38q3gtpp5ypz09jrd8p993snjwnm68cph4ftwp22le34xd4r8ftspwshxhmnsdqqxqyjw5qcqpxsp5htlg8ydpywvsa7h3u4hdn77ehs4z4e844em0apjyvmqfkzqhhd2q9qgsqqqyssqszpxzxt9uuqzymr7zxcdccj5g69s8q7zzjs7sgxn9ejhnvdh6gqjcy22mss2yexunagm5r2gqczh8k24cwrqml3njskm548aruhpwssq9nvrvz";
         let invoice = crate::invoice::parse_invoice(invoice_str)?;
